requirements.txt
mongoengine
django-angular
djangorestframework
django-rest-framework-mongoengine
django-filter
djangorestframework-filters
django-crispy-forms
django-lazysignup
